Add files via upload
Amazon Review Scraper and Google Sheets Integration

You Should add google sheet API from google cloud for running code!

Description:

This project is a web scraper and automation tool designed to extract product reviews from Amazon and save the collected data in a Google Sheet. This application is built using Python and utilizes a combination of several libraries such as BeautifulSoup, Selenium, and Google APIs to create an efficient and user-friendly tool.

Key Features:

Web Scraping: The application scrapes Amazon product reviews using Selenium and BeautifulSoup libraries. It navigates through the product page, identifies the 'See all reviews' link, and extracts relevant information such as the username, star rating, review title, and review text.

Pagination Handling: The scraper handles pagination efficiently, iterating through all available review pages for the given product and collecting data from each page.

Data Formatting: The review text is formatted using the textwrap library to ensure it fits well within the columns of the Google Sheet.

Google Sheets API Integration: The extracted data is saved in a new Google Sheet created by the application. The Google Sheets API is used to create, share, and modify the spreadsheet.

Column Width Adjustment: The application automatically adjusts the column widths in the Google Sheet based on the length of the data, making it easier to read and analyze the collected reviews.

Sharing Functionality: The application shares the created Google Sheet with a specified email address, enabling seamless collaboration.

How to Use:

Input the Amazon product URL.
Provide a desired title for the Google Sheet.
Enter the email address to share the created Google Sheet with.
The application will scrape the reviews, create a Google Sheet, and store the data.
Access the shared Google Sheet to view and analyze the collected reviews.
This project demonstrates proficiency in web scraping, data extraction, data manipulation, and API integration. The application provides a practical solution for gathering product review data and storing it in an easily accessible and shareable format. This project can be a valuable addition to a portfolio, showcasing technical skills in Python, web scraping, and working with APIs.